for heroku

+++ db/migrate/20150704133201_drop_people.rb
@@ -1,5 +1,7 @@
 class DropPeople < ActiveRecord::Migration
   def change
-  	drop_table :people
+    if ActiveRecord::Base.connection.table_exist 'people'
+  	  drop_table :people
+  	end
   end
 end
