+++ db/migrate/20150905184943_drop_table_users.rb
@@ -1,5 +1,7 @@
 class DropTableUsers < ActiveRecord::Migration
   def change
-    drop_table :users
+    if ActiveRecord::Base.connection.table_exist 'users'
+      drop_table :users
+    end
   end
 end
